Add checkDuplicate.js module
