+++ index.js
@@ -1,19 +1,6 @@
 import * as bookC from '/modules/bookClass.js';
 import * as constE  from '/modules/constantElement.js';
-
-const CheckDuplicate = function (title, author) {
-  const b = bookC.Books.books.filter(
-    (book) => title === book.title && author === book.author,
-  );
-  if (b.length !== 0) {
-    constE.errorMsg.classList.remove('hidden');
-    constE.successMsg.classList.add('hidden');
-    return true;
-  }
-  constE.errorMsg.classList.add('hidden');
-  constE.successMsg.classList.remove('hidden');
-  return false;
-};
+import * as checkD from '/modules/checkDuplicate.js';
 
 let id;
 // Add: when I click on Add button
@@ -24,7 +11,7 @@ constE.addBtn.addEventListener('submit', (e) => {
   const title = constE.titleInput.value;
   const author = constE.authorInput.value;
 
-  if (CheckDuplicate(title, author)) return;
+  if (checkD.CheckDuplicate(title, author)) return;
 
   const newBook = new bookC.Books(id, title, author);
   newBook.addBook();
